test: move regression test to its own file
hot.test.ts has a pre-existing debug-only failure
('should not hot reload when a random file is written') on main,
so the new test can't be gated cleanly there.
